Fixed AppearingDialogueControllerTests being in the wrong namespace and missing a letter in the name
